Moved Consumer and Broker to separate file
